requirements.txt
streamlit
streamlit-option-menu
paho-mqtt